refactor: split ask projection test contracts
